CUDA check before importing xFormers.
Checks if torch was compiled with CUDA before attempting to import
xFormers. Regardless of whether CUDA is available or not, torch needs to
be compiled with CUDA in order for tensors to be mappable to device, so
if it's not, attempting to import xFormers might succeed if the user
has xFormers installed in their environment.